requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
boto3==1.34.26
botocore==1.34.26
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
cryptography==42.0.0
databases==0.8.0
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jose==1.0.0
jwt==1.3.1
MarkupSafe==2.1.4
Naked==0.1.32
orjson==3.9.12
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
starlette==0.35.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
